# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Db server API description.",
        "title": "Db server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://openitstudio.ru",
            "email": "fedor@support-pc.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/admin/auth": {
            "post": {
                "description": "Authenticate in admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Login to admin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email for login",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password for login",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
...
        "/admin/users": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "List users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.User"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Create user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Admin"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
        "/admin/users/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "User detail info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Admin"
                ],
                "summary": "User info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Admin"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    },
                    {
                        "type": "string",
                        "description": "User info",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Delete user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Admin"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/device/register": {
            "post": {
                "description": "Register device to receive push",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Push messages",
                    "Public Api"
                ],
                "summary": "Register device",
                "parameters": [
                    {
                        "description": "Device info",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserDevice"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.UserDevice"
                        }
                    }
                }
            }
        },
...
        "/api/user/auth": {
            "post": {
                "description": "Authenticate via api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Public Api"
                ],
                "summary": "Login via api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email for login",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password for login",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
        "/api/user/me": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Get current user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Public Api"
                ],
                "summary": "User info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email for login",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password for login",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
...
        "/api/user/register": {
            "post": {
                "description": "Register via api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Public Api"
                ],
                "summary": "Register via api",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email for login",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Password for login",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user.CreateUserForm": {
            "type": "object",
            "properties": {
                "Email": {
                    "description": "new User email",
                    "type": "string"
                },
                "Password": {
                    "description": "new User password",
                    "type": "string"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Is user active",
                    "type": "boolean"
                },
                "admin": {
                    "description": "Is user admin",
                    "type": "boolean"
                },
                "created_at": {
                    "description": "Created at date time",
                    "type": "string"
                },
                "devices": {
                    "description": "user devices",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.UserDevice"
                    }
                },
                "email": {
                    "description": "User email",
                    "type": "string"
                },
                "id": {
                    "description": "The user UUID\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "last_login": {
                    "description": "Last login date time",
                    "type": "string"
                },
                "token": {
                    "description": "Auth token",
                    "type": "string"
                }
            }
        },
        "user.UserDevice": {
            "type": "object",
            "properties": {
                "device": {
                    "description": "Device type can be android | ios | macos | windows | linux | web\nexample: 'linux'",
                    "type": "string"
                },
                "device_token": {
                    "description": "UUID Device Token\n9f80fcc8-0102-4795-94c0-4190c168ffc2",
                    "type": "string"
                },
                "id": {
                    "description": "The UUID of a device\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "user_id": {
                    "description": "The UUID of owned user\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}